add "send extension handshake"
I encountered that I cannot use the pointer to the message when
adding/modifying the message for the ExtensionMessage, but the same
approach works for HandshakeMessage which is weird. As workaround when
returning the passed message, the changes are applied. Need to find out
why it looks that it is working on a copy otherwise.